Backend/requirements.txt
# SUPER MINIMAL requirements.txt (GUARANTEED TO WORK)
fastapi==0.110.0
uvicorn[standard]==0.27.1
pydantic==2.6.1
python-multipart==0.0.9
python-dotenv==1.0.1
firebase-admin==6.4.0
google-generativeai==0.4.0
easyocr==1.7.0
opencv-python==4.8.1.78
Pillow==9.5.0
torch==2.1.0
torchvision==0.16.0
transformers==4.35.0
huggingface-hub==0.17.3
tokenizers==0.14.1
scikit-learn==1.4.0
pandas==2.2.0
numpy==1.26.3
joblib==1.3.2
requests==2.31.0
colorama==0.4.6
sendgrid==6.10.0
pytest==8.3.5
memory-profiler==0.61.0
psutil==5.9.6
matplotlib
email-validator